Fixed Player animation state machine and script. Created new camera prefab.
Building script for camera. Currently the camera can be bound within a level and still follow the player. The UI is dynamically generated as a child of the camera game object. This script needs debugging. Roll charges also implemented. Also added the food object which randomly is assigned a sprite on spawn.